sw/pi: Properly use 0xFFFF0000 as a magic for PI Writes
The N64 may set the LSB of an address when performing
unaligned DMA from the PI. This leads to issues since
we used the LSB as a write command.

Instead of just checking the LSB, we can check the 16 MSB
of the value to be 0xFFFF0000.

This shouldn't be a problem since performing PI read/writes
on >= 0x80000000 isn't supported anyway.